Issue #377 - use the VMUSB drivers in place rather than as copies
Use MDPP32 with SCP firmware support directly from vmusb/daqconfig
