+++ main/mvlc/Makefile.am
@@ -69,11 +69,12 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CMarker.h $(IMPORT_DIR)/CMarker.cpp \
 	$(IMPORT_DIR)/ CMDPP.h \
 	$(IMPORT_DIR)/CMDPP16QDC.cpp  $(IMPORT_DIR)/CMDPP16QDC.h \
-	$(IMPORT_DIR)/CMDPP32QDC.cpp  	$(IMPORT_DIR)/CMDPP32QDC.h 
+	$(IMPORT_DIR)/CMDPP32QDC.cpp  	$(IMPORT_DIR)/CMDPP32QDC.h \
+	$(IMPORT_DIR)/CMDPP32SCP.cpp  $(IMPORT_DIR)/CMDPP32SCP.h
 
 LOCAL_MODULES= 	MADC32Registers.h \
 	CMTDC32.h CMTDC32.cpp \
-	CMDPP32SCP.cpp  CMDPP32SCP.h   CMDPPQDC.h \
+	    CMDPPQDC.h \
 	CNADC2530.h CNADC2530.cpp CV977.h CV977.cpp \
 	CV1495sc.h CV1495sc.cpp CV1729.h CV1729.cpp \
 	CCAENV1x90Opcodes.h CCAENV1x90Registers.h CV1x90.cpp CV1x90.h \
